+ RFE: allow SO to set the weightclass of a unit based on build table settings + RFE: allow force size to face to use float

+++ src/client/campaign/CArmy.java
@@ -82,7 +82,7 @@ public class CArmy extends Army {
 		}
 		
 		if ( ST.hasMoreTokens() )
-			this.setOpForceSize(Integer.parseInt(ST.nextToken()));
+			this.setOpForceSize(Float.parseFloat(ST.nextToken()));
 		
 		if ( ST.hasMoreElements())
 		{
